Fix centering of shorttitle in footer
Up to now, the alignment was done using \hfills, which will cause the
shorttitle to be centered between author and pagenumber, but not
centered with regards to the current page.
This commit introduces a tikz overlay instead. Please note: for the
standout-style, the bar has to be drawn twice: This is the only way to
draw the background and text on the main layer, without resorting to
pgfonlayer, which might have the unfortunate result of mixing with
included tikz graphics.